Create README.md
For various reasons, some programs don't have Russian localization. Sometimes the Russian language is removed for different reasons—an author's whim is a strange thing, but they are the author, so there's not much point in arguing.

That's why I decided to collect my translations here, which I'll probably be able to update if there's anything new.

If someone downloads a program and discovers it lacks a Russian interface, they can find the translation file in this spot.

I don't have many translations. First and foremost, I'm passionate about emulating game consoles. The very first translation I attempted (and still maintain) was for the "Project64 1.6.2" emulator. Later, I started trying other projects, updating the translation for the original Project64.

Then I translated the best HLE video plugin, "GLideN64." Interestingly, the project's author is Russian-speaking, and the video plugin had all sorts of languages except Russian—now it does.

After that, I translated the "RMG" emulator because it wasn't translated and its interface is incredibly good. If the author adds language support, you'll know where to get the translation.

All my translations might be incomplete, as some localization files simply lack certain text strings. Only what's present in the translation file has been translated.

As for the quality of the translations, I can't claim to be the best in the world, but I can say they're at least acceptable, which is enough for normal use, or perhaps even more than acceptable ;)

In any case, I'm doing this for a small number of people, and I know it. In Russia, not many people are into retro gaming, especially in the context of Nintendo. So, these translations are primarily for myself, and for the few others who might find them useful.

Most importantly, remember that no one is forcing you to use them. The work I do is a hobby in my spare time. If someone finds it interesting, I'm fine with that; I'm not asking for anything in return.

I wish everyone all the best. If you find something in a translation that needs to be changed, and you're absolutely sure you're right, I'll probably listen. Let me know if there's anything, but provide arguments, not just because you feel like it or it seems right to you.

Peace and love to all, "SatanNintendo."